Refactor main page from 1200+ to 200+ lines
